button working now

+++ screen3.php
@@ -47,6 +47,9 @@
 		<?
 		if(array_key_exists('button1', $_POST)) { 			//call function
 			echo 'Hello!';
+			$add = "insert into order_book (order_number, book_isbn, quantity) values(1, ".$ISBN.", 1);";
+			echo $add;
+			pg_query($db, $add);
 			addToCart(); 
 			} 
 		$db = pg_connect("host=ec2-3-218-75-21.compute-1.amazonaws.com dbname=d8p0qs8v3fbf9m user=gymsvpkhkckshh password=68db7ff943798b07abc442d46449c9d2f4bfcd38be0f79023a630bf67b3b3a8a");
@@ -65,6 +68,14 @@
 			$Price = $row[4];
 			$book = "<td rowspan='2' align='left'>".$Title."</br>".$Author."</br><b>Publisher:</b>".$Publisher.",</br><b>ISBN:</b>".$ISBN."</t> <b>Price:</b>".$Price."</td>";
 			$review = "<input name='review' id='review' type='submit' value='".$ISBN."' onClick='review(".$ISBN.", ".$Title.")'></input>";
+
+			if(array_key_exists('button1', $_POST)) { 			//call function
+				echo 'Hello Again!';
+				$add = "insert into order_book (order_number, book_isbn, quantity) values(1, ".$ISBN.", 1);";
+				echo $add;
+				pg_query($db, $add);
+				addToCart(); 
+				} 
 					echo '<tr>';
 						echo "<td align='left'>";
 						echo '<form method="post">';
